v1.2.0 to fix bug
### Releases v1.2.0

1. Permit using servos with different pulse ranges simultaneously. Check [Fix Bug #5 : Permit using servos with different pulse ranges simultaneously #4](https://github.com/khoih-prog/RP2040_ISR_Servo/pull/4)
2. Delete left-over `cpp`

+++ examples/nRF52_MultipleRandomServos/nRF52_MultipleRandomServos.ino
@@ -52,7 +52,6 @@
 #define TIMER_INTERRUPT_DEBUG       4
 #define ISR_SERVO_DEBUG             4
 
-// To be included only in main(), .ino with setup() to avoid `Multiple Definitions` Linker Error
 #include "NRF52_ISR_Servo.h"
 
 // Published values for SG90 servos; adjust if needed
@@ -68,7 +67,7 @@
 
 typedef struct
 {
-  int8_t  servoIndex;
+  int     servoIndex;
   uint8_t servoPin;
 } ISR_servo_t;
